# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FoodHelper API
  version: v1
  description: API for managing products, recipes, and nutritional calculations.
servers:
  - url: /
    description: Current server
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/money-boxes:
    get:
      tags:
        - Money boxes
      summary: List all money boxes
      description: Returns user-owned and manually created money boxes in one list, including balances and movements.
      operationId: listMoneyBoxes
      responses:
        '200':
          description: Money boxes returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MoneyBoxResponse'
    post:
      tags:
        - Money boxes
      summary: Create manual money box
      description: Creates a money box that is not owned by a user.
      operationId: createManualMoneyBox
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMoneyBoxRequest'
            examples:
              default:
                value:
                  name: Household cash
      responses:
        '201':
          description: Manual money box created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoneyBoxResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Manual money box name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/money-boxes/{moneyBoxId}:
    get:
      tags:
        - Money boxes
      summary: Get money box
      description: Returns one user-owned or manual money box by identifier.
      operationId: getMoneyBox
      parameters:
        - name: moneyBoxId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Money box returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoneyBoxResponse'
        '404':
          description: Money box not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags:
        - Money boxes
      summary: Delete manual money box
      description: Deletes a manual money box and all its movements in one transaction. User-owned money boxes cannot be deleted.
      operationId: deleteManualMoneyBox
      parameters:
        - name: moneyBoxId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Manual money box deleted
        '404':
          description: Money box not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: User-owned money box cannot be deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/money-boxes/{moneyBoxId}/movements:
    post:
      tags:
        - Money boxes
      summary: Add money box movement
      description: Adds a signed movement to a user-owned or manual money box.
      operationId: addMoneyBoxMovement
      parameters:
        - name: moneyBoxId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserMoneyMovementRequest'
      responses:
        '201':
          description: Money movement created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoneyBoxMovementResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Money box not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/money-boxes/{moneyBoxId}/movements/{movementId}:
    delete:
      tags:
        - Money boxes
      summary: Delete money box movement
      description: Deletes a movement only when it belongs to the selected money box and is not linked to a menu.
      operationId: deleteMoneyBoxMovement
      parameters:
        - name: moneyBoxId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: movementId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Money movement deleted
        '404':
          description: Money box movement not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Money movement is linked to a menu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
    ApiError:
      type: object
      properties:
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Invalid request body
        path:
          type: string
          example: /api/v1/products
        timestamp:
          type: string
          format: date-time
...
    CreateMoneyBoxRequest:
      description: Payload for creating a manual money box.
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 150
          description: Unique manual money box name
          example: Household cash
    MoneyBoxMovementResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Movement identifier
          example: 1
        moneyBoxId:
          type: integer
          format: int64
          description: Money box identifier
          example: 3
        userId:
          type:
            - integer
            - 'null'
          format: int64
          description: Owning user identifier for user money boxes; null for manual money boxes
          example: 1
        amount:
          type: number
          description: Signed movement amount
          example: -12.50
        description:
          type: string
          description: Movement description
          example: Weekly groceries
        menuId:
          type:
            - integer
            - 'null'
          format: int64
          description: Related menu identifier when generated by a menu cost
          example: 7
        createdAt:
          type: string
          format: date-time
          description: UTC instant when the movement was created
    MoneyBoxResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Money box identifier
          example: 3
        type:
          type: string
          enum:
            - USER
            - MANUAL
          description: Money box origin
          example: MANUAL
        name:
          type: string
          description: Display name
          example: Household cash
        userId:
          type:
            - integer
            - 'null'
          format: int64
          description: Owning user identifier; null for manual money boxes
          example: 1
        username:
          type:
            - string
            - 'null'
          description: Owning username; null for manual money boxes
          example: elias
        balance:
          type: number
          description: Current balance calculated from all signed movements
          example: 87.50
        movements:
          type: array
          description: Movements ordered from newest to oldest
          items:
            $ref: '#/components/schemas/MoneyBoxMovementResponse'
    CreateUserMoneyMovementRequest:
      description: Signed money movement to add to a user's money box.
      type: object
      required:
        - amount
      properties:
        amount:
          type: number
          description: Signed amount. Positive values add money and negative values subtract money.
          example: -12.50
        description:
          type: string
          description: Optional movement description
          example: Weekly groceries
...