# src/main/java/com/eliascanalesnieto/foodhelper/presentation/MoneyBoxMovementResponse.java
package com.eliascanalesnieto.foodhelper.presentation;

import io.swagger.v3.oas.annotations.media.Schema;
import java.math.BigDecimal;
import java.time.Instant;

@Schema(name = "MoneyBoxMovementResponse", description = "Signed movement registered in a money box")
public record MoneyBoxMovementResponse(
        @Schema(description = "Movement identifier", example = "1") Long id,
        @Schema(description = "Money box identifier", example = "3") Long moneyBoxId,
        @Schema(description = "Owning user identifier for user money boxes; null for manual money boxes", example = "1") Long userId,
        @Schema(description = "Signed movement amount", example = "-12.50") BigDecimal amount,
        @Schema(description = "Movement description", example = "Weekly groceries") String description,
        @Schema(description = "Related menu identifier when generated by a menu cost", example = "7") Long menuId,
        @Schema(description = "UTC instant when the movement was created") Instant createdAt
) {
}